+++ voucher-impl/src/test/java/com/dwarfeng/voucher/impl/service/CheckerSupportMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -57,6 +58,9 @@ public class CheckerSupportMaintainServiceImplTest {
             }
         } finally {
             for (CheckerSupport checkerSupport : checkerSupports) {
+                if (Objects.isNull(checkerSupport.getKey())) {
+                    continue;
+                }
                 checkerSupportMaintainService.deleteIfExists(checkerSupport.getKey());
             }
         }
